==> Part2.ipynb <==
# Сессия 2

## ВВЕДЕНИЕ:
На этой сессии, я буду продолжать исследования, опираясь на то, что я уже разработал. Мне необходимо выполнить предварительную обработку данных, найти ключевые слова, биграммы/триграммы для их дальнейшего использования в построении моделей. 


---

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import regex as re

Чтение документа

df = pd.read_csv("df.csv")

Реиндексация

df.reset_index(drop=True, inplace=True)

df

### 2.1. Предварительная обработка данных 

Импорт необходимых библиотек

import nltk
import string
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
import pymorphy2

Для тестов вытащу одну жалобу

#Использовал iloc что бы вытащить 0 строку по индексу
sentence_example = df.iloc[0]["appeal"]

Обьявление констант для функции которая описана ниже

#MorphAnalyzer нужен для преведения слова в начальную фонрму (леметизация)
morph = pymorphy2.MorphAnalyzer()
#stopwords нудна для удаления предлогов и прочего мусора который создает шум и ухудшает качество алгоритмов
russian_stop_words = stopwords.words("russian")

Теперь нужна функция, которую потом можно будет сунуть в apply

#Функция будет принимтаь на вход строку и будевое значение (может возникнуть ситуация в которой будут не нужны стоп слова)
def tokinize_sentence(sentense: str, remove_stop_word: bool = True):
    #Первый этап - разделения предложения на токены или же отдельные кусочки
    tokens = word_tokenize(sentense, language="russian")
    #Далее необходимо убрать пунктуацию тк как они не нужна и создает много шумов
    tokens = [i for i in tokens if i not in string.punctuation]
    #Теперь удаление стоп слов
    if remove_stop_word:
        tokens = [i for i in tokens if i not in russian_stop_words]
    
    #В жалобах появляется много непонятных Х по этому от них тоже необходимо избавится
    stammed_tokens_origin = []
    for i in tokens:
        #запишу в отдельный список все слова которые не содержат Х
        if 'X' not in i:
            stammed_tokens_origin.append(i)
    #Переприсвоил список для удобства
    tokens = stammed_tokens_origin
    
    #А теперь вернем словам их первоначальную форму
    tokens = [morph.parse(i)[0].normal_form for i in tokens]
    
    return tokens

Удалю все спец символы и цифры, что бы остались только слова

#Удалю все спец символы и цифры при помощи регулярных выражений, а strip удалит лишние пробелы в начале и в конце предложения
df["clean_appeal"] = df["appeal"].apply(lambda x: re.compile('[^а-яА-Я ]').sub('', x)) 
#Теперь применю вышеописанную функцию к каждому элементу атрибьута appeal
df["clean_appeal"] = df["clean_appeal"].apply(lambda x: tokinize_sentence(x, remove_stop_word=True))

Решил так же удалить все значения списка у которых менее 3х символов (сложно догадаться что имелось ввиду когда видешь "пра" или "й")

#Удаление огрызков слов
df["clean_appeal"] = df["clean_appeal"].apply(lambda x: [i for i in x if len(i) > 4])

Теперь соеденю все слова в предложения и помещу в отдельный атрибут

#Для этой зажачи вновь использую метод apply
df["clean_sentence"] = df["clean_appeal"].apply(lambda x:' '.join(x))

df

### 2.2. Поиск ключевых слов/биграмм/триграмм

## Алгоритмы нахождения ключевых слов
Рассмотрим несколько алгоритмов для нахождения ключевых слов
### TfidfVectorizer
Преобразует набор необработанных документов в матрицу функций TF-IDF
### HashingVectorizer
Преобразует коллекции текстовых документов в матрицу вхождений токенов
### CountVectorizer
Преобразует коллекцию текстовых документов в матрицу счетчиков токенов

Обоснование выбора алгоритма: я выбираю алгоритм `TfidfVectorizer`, поскольку он способен определять важность слов или ngram в векторном формате разреженной матрицы, что способно повысить качество определения важности(они же ключевые слова или ngrma'ы) таких ngram

### Образование новых признаков с ключевыми словами
Ниже представленны методы, которые формируют атрибуты с unigram, bigram и trigram. Также есть возможность получения лучшей ngramm в поле с помощью метрики tf-idf в словаре поля, если это необходимо 

from sklearn.feature_extraction.text import TfidfVectorizer

vectorizer = TfidfVectorizer(min_df=15, max_df=0.95, ngram_range=(1,1))
X = vectorizer.fit_transform(df['clean_sentence'])
vect_unigram = pd.DataFrame(X.todense(), columns=vectorizer.get_feature_names())
vect_unigram.head()

Биграмм

vectorizer = TfidfVectorizer(min_df=15, max_df=0.95, ngram_range=(1,2))
X = vectorizer.fit_transform(df['clean_sentence'])
vect_begram = pd.DataFrame(X.todense(), columns=vectorizer.get_feature_names())
vect_begram.head()

Триграмм

vectorizer = TfidfVectorizer(min_df=15, max_df=0.95, ngram_range=(1,3))
X = vectorizer.fit_transform(df['clean_sentence'])
vect_trigram = pd.DataFrame(X.todense(), columns=vectorizer.get_feature_names())
vect_trigram.head()

df2 = pd.concat([df, vect_unigram, vect_begram, vect_trigram], axis=1)

df2

df2.isna().sum()

### 2.3. Подготовка отчета

#### В ходе работы было выполнено:
* 2.1. Предварительная обработка данных - Обработка данных с помощью методов NLP: Токенизация; Удаление стоп-слов; Удаление пунктуации; Удаление спецсимволов; Лемматизация;
* 2.2. Поиск ключевых слов/биграмм/триграмм - Бало проаналищированно 3 алгоритма векторизации (TfidfVectorizer, HashingVectorizer, CountVectorizer); Были простроенны нграммы по алгоритму векторизации TfidfVectorizer;

#Сохранение набора данных
df.to_csv("df.csv", encoding='utf-8-sig', index=False)
df2.to_csv("result_data_clean.csv", encoding='utf-8-sig', index=False)
