==> Part4.ipynb <==
# Сессия 4

---

import pandas as pd
import warnings
warnings.filterwarnings("ignore")

Чтение данных из csv файла

df = pd.read_csv("ndata.csv", encoding='utf-8-sig')

df

## 4.1. Разбиение набора данных на обучающую и тестовую выборки

Разобью данные 67% на 33% так как сичтаю это достаточным для обучения и тестирования

from sklearn.model_selection import train_test_split
#Использовал параметр stratify для того что бы какждый вид кластера присутсвтовал в обучающей и тестовой выборке
X_train, X_test, y_train, y_test = train_test_split(df[['district_name', 'sidebar.author_name','clean_sentence']], df['KMeans Clus Label'], test_size=0.33, random_state=42, stratify=df['KMeans Clus Label'])

## 4.2. Создание дополнительных признаков

Добавлю атрибут который будет содержать количество слов из очещеного предложения

#Использую метод len для подщета количетсва элементов в списке который получу от метода split
df["word_count"] = df['clean_sentence'].apply(lambda x: len(x.split(" ")))

Еще добавлю атрибут в котором будет находится средняя длинна слова в очищеном предложении

#Функция для подщета средней длины слова
def avg_word(sentence):
    words = sentence.split()
    return (sum(len(word) for word in words)/len(words))
#применяю функцию через apply
df['avg_word'] = df['clean_sentence'].apply(lambda x: avg_word(x))

## 4.3. Классификация обращений граждан по категориям обращений

Для обьедененя разнотипных атрибутов воспользуюсь модулем Pipeline

####  Добавление атрибутьов будет происходить при помощи модуля Pipeline

from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.model_selection import RandomizedSearchCV
from sklearn.pipeline import Pipeline, FeatureUnion
from sklearn import metrics

# text and numeric classes that use sklearn base libaries
class TextTransformer(BaseEstimator, TransformerMixin):
    """
    Transform text features
    """
    def __init__(self, key):
        self.key = key

    def fit(self, X, y=None, *parg, **kwarg):
        return self

    def transform(self, X):
        return X[self.key]
    
class NumberTransformer(BaseEstimator, TransformerMixin):
    """
    Transform numeric features
    """
    def __init__(self, key):
        self.key = key

    def fit(self, X, y=None):
        return self

    def transform(self, X):
        return X[[self.key]]


# use the term-frequency inverse document frequency vectorizer to transfrom count of text
# into a weighed matrix of term importance
vec_tdidf = TfidfVectorizer(ngram_range=(1,1), analyzer='word', norm='l2')

# compile both the TextTransformer and TfidfVectorizer 
# to the text 'Text_Feature' 
text_sentence = Pipeline([
                ('transformer', TextTransformer(key='clean_sentence')),
                ('vectorizer', vec_tdidf)
                ])

# compile the NumberTransformer to 'text_numeric'
# and 'text_numeric2' numeric features
text_numeric = Pipeline([
                ('transformer', NumberTransformer(key='word_count')),
                ])

text_numeric2 = Pipeline([
                ('transformer', NumberTransformer(key='avg_word')),
                ])



# combine all of the features, text and numeric together
features = FeatureUnion([('text_attribute', text_sentence),
                      ('word_count', text_numeric),
                      ('avg_word', text_numeric2),
                      ])


text_numeric_features = ['clean_sentence', 'word_count', 'avg_word']
predictor = 'reason.category.name'

X_train, X_test, y_train, y_test = train_test_split(df[text_numeric_features], df[predictor], 
                                                    test_size=0.25, random_state=42, stratify=df[predictor])

#### для задачи классификации решил взять алгоритмы RandomForestClassifier, KNN и AdaBoost так как они хорошо подходят для обработки нескольких признаков и являются довольно быстрыми, что упрастит подбор гипер параметров

### RandomForestClassifier

Настройка гипер параметров

# Minimum number of samples per leaf
n_estimators = [99, 200]

# Minimum number of samples to split a node
max_depth = [25, 50]

hyperparameter_grid = {'n_estimators': n_estimators,
                       'max_depth': max_depth}

RandomizedSearchCV это инструмент, который позволяет сделать прогон циклом по нескольким гиперпараметрам и выбрать лучший

from sklearn.ensemble import RandomForestClassifier

# Create the model to use for hyperparameter tuning
model = RandomForestClassifier()

# Set up the random search with 4-fold cross validation
clf = RandomizedSearchCV(estimator=model,
                               param_distributions=hyperparameter_grid,
                               cv=3, n_iter=20, 
                               scoring = 'neg_mean_absolute_error',
                               n_jobs = -1, verbose = 1, 
                               return_train_score = True,
                               random_state=42)

# unite the features and classfier together
pipe = Pipeline([('features', features),
                 ('clf',clf)
                 ])

Обучение модели

# fit the model
pipe.fit(X_train, y_train)

# unite the features and classfier together
pipeforst2 = Pipeline([('features', features),
                 ('clf',clf.best_estimator_)
                 ])

# fit the model
pipeforst2.fit(X_train, y_train)

# predict from the test set
preds = pipeforst2.predict(X_test)

# print out your accuracy
from sklearn.metrics import classification_report
print("best was - ", clf.best_estimator_)
print(classification_report(y_test, preds))

метрики - weighted accuracy macro по F1-score показывают не плохой результат.

### KNN

Настройка гипер параметров

# Minimum number of samples per leaf
n_neighbors = [4, 7]

# Minimum number of samples to split a node
leaf_size = [29, 69]

hyperparameter_grid = {'n_neighbors': n_neighbors,
                       'leaf_size': leaf_size}

Так же использую RandomizedSearchCV для подбора нескольких параметров

from sklearn.neighbors import KNeighborsClassifier
# Create the model to use for hyperparameter tuning
model = KNeighborsClassifier()

# Set up the random search with 4-fold cross validation
clf = RandomizedSearchCV(estimator=model,
                               param_distributions=hyperparameter_grid,
                               cv=3, n_iter=20, 
                               scoring = 'neg_mean_absolute_error',
                               n_jobs = -1, verbose = 1, 
                               return_train_score = True,
                               random_state=42)

# unite the features and classfier together
pipe = Pipeline([('features', features),
                 ('clf',clf)
                 ])

Обучение модели

# fit the model
pipe.fit(X_train, y_train)

# unite the features and classfier together
pipe2 = Pipeline([('features', features),
                 ('clf',clf.best_estimator_)
                 ])

# fit the model
pipe2.fit(X_train, y_train)

# predict from the test set
preds = pipe2.predict(X_test)

# print out your accuracy
print("best was - ", clf.best_estimator_)
print(classification_report(y_test, preds))

На удивление KNN справился хуже с этой задачей

### AdaBoost

Настройка гипер параметров

# Minimum number of samples per leaf
n_estimators = [49, 100]

# Minimum number of samples to split a node
learning_rate = [1.1, 3]

hyperparameter_grid = {'n_estimators': n_estimators,
                       'learning_rate': learning_rate}

Так же использую RandomizedSearchCV для подбора нескольких параметров

from sklearn.ensemble import AdaBoostClassifier
# Create the model to use for hyperparameter tuning
model = AdaBoostClassifier()

# Set up the random search with 4-fold cross validation
clf = RandomizedSearchCV(estimator=model,
                               param_distributions=hyperparameter_grid,
                               cv=3, n_iter=20, 
                               scoring = 'neg_mean_absolute_error',
                               n_jobs = -1, verbose = 1, 
                               return_train_score = True,
                               random_state=42)

# unite the features and classfier together
pipe = Pipeline([('features', features),
                 ('clf',clf)
                 ])

Обучение модели

# fit the model
pipe.fit(X_train, y_train)

# unite the features and classfier together
pipe2 = Pipeline([('features', features),
                 ('clf',clf.best_estimator_)
                 ])

# fit the model
pipe2.fit(X_train, y_train)

# predict from the test set
preds = pipe2.predict(X_test)

# print out your accuracy
print("best was - ", clf.best_estimator_)
print(classification_report(y_test, preds))

AdaBoost показал самый плохой результат по итогам метрик

#### Лучшим из этих алгормитмов классификаци для данной задачи является  RandomForestClassifier

### 4.4. Подготовка отчета

df

#### Было выполнено:
* 4.1. Разбиение набора данных на обучающую и тестовую выборки - Выборка была разбита на тестовую и обучающую со сбалансированым разделением и стратификацией по виду кластеров;
* 4.2. Создание дополнительных признаков - Было сформировано 2 атрибута (word_count и avg_word);
* 4.3. Классификация обращений граждан по категориям обращений - Была выполнена классификация несколькими способами (RandomForestClassifier, KNN и AdaBoost); Была выполнена нстройка гипер параметров при помощи RandomizedSearchCV; Оценка качества осуществлялась по средствам метрик weighted accuracy macro по F1-score; По результатам метрик был выбран лучший алгоритм (RandomForestClassifier);

df.to_csv("nddata.csv", index=False, encoding="utf-8-sig")

import pickle
pickle.dump(model, open("pickle.dat", "wb"))

---

df2 = df[df["reason.category.name"] == "Благоустройство"][::14]
df3 = df[df["reason.category.name"] == "Содержание МКД"][::12]
df = df[df["reason.category.name"] != "Благоустройство"]
df = df[df["reason.category.name"] != "Содержание МКД"]
df = pd.concat([df, df2, df3])
df.reset_index(drop=True, inplace=True)
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer
tfidf = TfidfVectorizer()
X3 = tfidf.fit_transform(df['clean_sentence'])
X_train, X_test, y_train, y_test = train_test_split(X3, df["reason.category.name"], test_size=0.33, random_state=42, stratify=df["reason.category.name"])
model = RandomForestClassifier()
model.fit(X_train, y_train)
preds = model.predict(X_test)
x = tfidf.transform(["необходимо отчистить стена парадный"])
model.predict(x)
import pickle
pickle.dump(model, open("pickle.dat", "wb"))
